info service moved into service package
